Update Glyph Client for license race condition with automatic Glyph Server
A race condition was occurring when port=0 where the Glyph Server subprocess was not quite ready to receive commands. This could result in zombie subprocesses which may consume a pointwise license.

Added a new optional keyword argument to specify the name of the program to run as the Glyph Server. The environment variable 'PWI-GLYPH_SERVER_TCLSH' is no longer used to specify a non-standard program name. This closes a potential security hole related to OS environment.