Update GlyphClient for platform-specific issues
Improved handling of port=0 subprocess all supported platforms.

Modified optional server launch argument to accept a list or tuple of values, where the first value is the program to run and remaining arguments, if present, are passed as arguments to the server program.

Added NoLicenseError exception class for cleaner client-side handling of license acquisition errors.

Added a check for port=0 server to be at least Pointwise V18.2.

Update PyPi version to 2.0.6.